Large changes: pdims now get unique names at init
This was necessary to move it closer to xr.DataArrays, which supply the
ability of named dimensions. In order to make this as general as
possible, it was best to implement this directly in ParamSpace.

While imposing some restrictions on the naming of parameter dimensions,
this change hopefully makes it easier to work with paramspace.