Make state map output proper xr.DataArrays
It is now possible to get a state_map with proper coordinates and an
active_state_map that only includes the active values and supports
location indexing.

To that end, a number of additions were necessary:
 - let the ParamDims and ParamSpace have coordinates
 - have certain restrictions on the values that ParamDims can take
 - add a number of properties
 - adjust and extend tests

Also: improved state_map creation by avoiding to set the state vector
but just calculating the state number with the same method.