Add tox environments using strict minimal dependency versions
These are executed in additional CI jobs and assert that the lower bound
is actually still reasonable to specify. Prior to this, all jobs were
using the latest versions of packages, thus possibly not failing when
implementing new features that would not work with the lowest specified
version.

The strict version requirements are defined via a pip install extra
which is automatically created; this assures that the version numbers
only need to be specified in one place.