Improve short-circuiting in recursive_replace and recursive_collect
This optimisation assumes that many objects in dicts are strings, which
are actually meant to be ignored for recursion. Thus, the check for
to-be-ignored types should be done first, such that the conditional is
short-circuited.

In absolute terms, evaluation of the recursion condition is fast.
However, as part of the recursion, this may be called many million
times, such that it is worth optimising.